+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 12/29/2023
+- Major: Added `check-in-checker` to check if user has checked in or not to avoid missing check-in
 - Minor: Add `getSignData` method to CheckIn class
 
 ## 12/28/2023